backend/requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
PyJWT>=2.0.0
python-multipart>=0.0.5
certifi>=2024.8.30
pdfplumber
python-docx
openai
fastapi-mail
